requirements.txt
streamlit==1.26.0
tensorflow>=2.12.0
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
scikit-learn==1.2.2
joblib==1.2.0
Pillow==9.5.0